InstaReel/recommendation-system/flask_app/requirements.txt
blinker
click
colorama
Flask
itsdangerous
Jinja2
numpy
pandas 
pymongo
python-dateutil
python-dotenv
Werkzeug
flask-cors
scikit-learn  
